tdf#140982 sw ChangesInMargin: fix annotation ranges
Annotations of tracked deletions collapsed at the
end of the annotation range in ChangesInMargin mode
(as before saving the document, see
commit d325cd0c69b7c0cc4f47105749a98995de81cc9d
"tdf#115815 sw: fix lost annotation ranges of redlines").

+++ sw/source/core/doc/docbm.cxx
@@ -1643,7 +1643,7 @@ namespace sw::mark
 
     // restore text ranges of annotations of tracked deletions
     // based on the helper bookmarks (which can survive I/O and hiding redlines)
-    void MarkManager::restoreAnnotationMarks()
+    void MarkManager::restoreAnnotationMarks(bool bDelete)
     {
         for (auto iter = getBookmarksBegin();
               iter != getBookmarksEnd(); )
@@ -1660,9 +1660,14 @@ namespace sw::mark
                     const SwPaM aPam((**iter).GetMarkStart(), (**pMark).GetMarkEnd());
                     repositionMark(*pMark, aPam);
                 }
-                deleteMark(&**iter);
-                // this invalidates iter, have to start over...
-                iter = getBookmarksBegin();
+                if (bDelete)
+                {
+                    deleteMark(&**iter);
+                    // this invalidates iter, have to start over...
+                    iter = getBookmarksBegin();
+                }
+                else
+                    ++iter;
             }
             else
                 ++iter;
